refactor: share windows process cleanup helpers
