+++ apps/backend/internal/agentctl/server/process/procattr_windows.go
@@ -6,15 +6,15 @@ import (
 	"errors"
 	"fmt"
 	"os/exec"
-	"strings"
 	"syscall"
-	"unsafe"
 
+	"github.com/kandev/kandev/internal/agentctl/server/winproc"
 	"golang.org/x/sys/windows"
 )
 
 // setProcGroup configures the command to run in its own process group.
-// On Windows, we use CREATE_NEW_PROCESS_GROUP flag.
+// On Windows, we also start suspended so installProcessLifecycle can bind
+// the process to a kill-on-close Job Object before it can spawn descendants.
 func setProcGroup(cmd *exec.Cmd) {
 	cmd.SysProcAttr = &syscall.SysProcAttr{
 		CreationFlags: syscall.CREATE_NEW_PROCESS_GROUP | windows.CREATE_SUSPENDED,
@@ -22,110 +22,19 @@ func setProcGroup(cmd *exec.Cmd) {
 }
 
 type processLifecycleHandle struct {
-	handle windows.Handle
+	job winproc.KillOnCloseJob
 }
 
 func installProcessLifecycle(cmd *exec.Cmd) (processLifecycleHandle, error) {
-	if cmd == nil || cmd.Process == nil {
-		return processLifecycleHandle{}, fmt.Errorf("process not started")
-	}
-	pid := cmd.Process.Pid
-	job, err := createKillOnCloseJob()
+	job, err := winproc.InstallKillOnCloseJobForSuspendedCommand(cmd)
 	if err != nil {
-		return processLifecycleHandle{}, errors.Join(err, resumeSuspendedProcess(pid))
-	}
-	procHandle, err := windows.OpenProcess(
-		windows.PROCESS_SET_QUOTA|windows.PROCESS_TERMINATE,
-		false,
-		uint32(pid),
-	)
-	if err != nil {
-		_ = windows.CloseHandle(job)
-		return processLifecycleHandle{}, errors.Join(
-			fmt.Errorf("OpenProcess(pid=%d): %w", pid, err),
-			resumeSuspendedProcess(pid),
-		)
-	}
-	defer windows.CloseHandle(procHandle)
-	if err := windows.AssignProcessToJobObject(job, procHandle); err != nil {
-		_ = windows.CloseHandle(job)
-		return processLifecycleHandle{}, errors.Join(
-			fmt.Errorf("AssignProcessToJobObject: %w", err),
-			resumeSuspendedProcess(pid),
-		)
-	}
-	if err := resumeSuspendedProcess(pid); err != nil {
-		_ = windows.CloseHandle(job)
 		return processLifecycleHandle{}, err
 	}
-	return processLifecycleHandle{handle: job}, nil
+	return processLifecycleHandle{job: job}, nil
 }
 
 func releaseProcessLifecycle(lifecycle processLifecycleHandle) {
-	if lifecycle.handle != 0 {
-		_ = windows.CloseHandle(lifecycle.handle)
-	}
-}
-
-func createKillOnCloseJob() (windows.Handle, error) {
-	job, err := windows.CreateJobObject(nil, nil)
-	if err != nil {
-		return 0, fmt.Errorf("CreateJobObject: %w", err)
-	}
-	info := windows.JOBOBJECT_EXTENDED_LIMIT_INFORMATION{
-		BasicLimitInformation: windows.JOBOBJECT_BASIC_LIMIT_INFORMATION{
-			LimitFlags: windows.JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE,
-		},
-	}
-	if _, err := windows.SetInformationJobObject(
-		job,
-		windows.JobObjectExtendedLimitInformation,
-		uintptr(unsafe.Pointer(&info)),
-		uint32(unsafe.Sizeof(info)),
-	); err != nil {
-		_ = windows.CloseHandle(job)
-		return 0, fmt.Errorf("SetInformationJobObject: %w", err)
-	}
-	return job, nil
-}
-
-func resumeSuspendedProcess(pid int) error {
-	snapshot, err := windows.CreateToolhelp32Snapshot(windows.TH32CS_SNAPTHREAD, 0)
-	if err != nil {
-		return fmt.Errorf("CreateToolhelp32Snapshot: %w", err)
-	}
-	defer windows.CloseHandle(snapshot)
-
-	entry := windows.ThreadEntry32{Size: uint32(unsafe.Sizeof(windows.ThreadEntry32{}))}
-	if err := windows.Thread32First(snapshot, &entry); err != nil {
-		return fmt.Errorf("Thread32First: %w", err)
-	}
-
-	resumed := 0
-	for {
-		if entry.OwnerProcessID == uint32(pid) {
-			thread, err := windows.OpenThread(windows.THREAD_SUSPEND_RESUME, false, entry.ThreadID)
-			if err != nil {
-				return fmt.Errorf("OpenThread(thread_id=%d): %w", entry.ThreadID, err)
-			}
-			if _, err := windows.ResumeThread(thread); err != nil {
-				_ = windows.CloseHandle(thread)
-				return fmt.Errorf("ResumeThread(thread_id=%d): %w", entry.ThreadID, err)
-			}
-			_ = windows.CloseHandle(thread)
-			resumed++
-		}
-		if err := windows.Thread32Next(snapshot, &entry); err != nil {
-			if errors.Is(err, windows.ERROR_NO_MORE_FILES) {
-				break
-			}
-			return fmt.Errorf("Thread32Next: %w", err)
-		}
-	}
-	if resumed == 0 {
-		return fmt.Errorf("no threads found for pid %d", pid)
-	}
-	return nil
+	_ = lifecycle.job.Close()
 }
 
 // killProcessGroup kills the entire process tree for the given PID.
@@ -152,26 +61,5 @@ func isProcessGroupMissing(err error) bool {
 }
 
 func runProcessTaskkill(args ...string) error {
-	output, err := exec.Command("taskkill", args...).CombinedOutput()
-	if err == nil {
-		return nil
-	}
-	msg := strings.TrimSpace(string(output))
-	if isProcessTaskkillMissing(msg) {
-		return syscall.ESRCH
-	}
-	if msg == "" {
-		return err
-	}
-	return fmt.Errorf("%w: %s", err, msg)
-}
-
-func isProcessTaskkillMissing(msg string) bool {
-	if msg == "" {
-		return false
-	}
-	lower := strings.ToLower(msg)
-	return strings.Contains(lower, "not found") ||
-		strings.Contains(lower, "not be found") ||
-		strings.Contains(lower, "no running instance")
+	return winproc.RunTaskkill(args...)
 }
