+++ apps/backend/internal/agentctl/server/utility/acp_command_windows.go
@@ -3,17 +3,18 @@
 package utility
 
 import (
-	"errors"
 	"fmt"
 	"os/exec"
 	"strconv"
 	"strings"
 	"syscall"
-	"unsafe"
 
+	"github.com/kandev/kandev/internal/agentctl/server/winproc"
 	"golang.org/x/sys/windows"
 )
 
+// setACPCommandProcAttr starts ACP utility commands suspended so the cleanup
+// lifecycle can attach a kill-on-close Job Object before the command runs.
 func setACPCommandProcAttr(cmd *exec.Cmd) {
 	cmd.SysProcAttr = &syscall.SysProcAttr{
 		CreationFlags: syscall.CREATE_NEW_PROCESS_GROUP | windows.CREATE_SUSPENDED,
@@ -21,110 +22,19 @@ func setACPCommandProcAttr(cmd *exec.Cmd) {
 }
 
 type acpCommandLifecycleHandle struct {
-	handle windows.Handle
+	job winproc.KillOnCloseJob
 }
 
 func installACPCommandLifecycle(cmd *exec.Cmd) (acpCommandLifecycleHandle, error) {
-	if cmd == nil || cmd.Process == nil {
-		return acpCommandLifecycleHandle{}, fmt.Errorf("process not started")
-	}
-	pid := cmd.Process.Pid
-	job, err := createACPKillOnCloseJob()
+	job, err := winproc.InstallKillOnCloseJobForSuspendedCommand(cmd)
 	if err != nil {
-		return acpCommandLifecycleHandle{}, errors.Join(err, resumeACPCommandProcess(pid))
-	}
-	procHandle, err := windows.OpenProcess(
-		windows.PROCESS_SET_QUOTA|windows.PROCESS_TERMINATE,
-		false,
-		uint32(pid),
-	)
-	if err != nil {
-		_ = windows.CloseHandle(job)
-		return acpCommandLifecycleHandle{}, errors.Join(
-			fmt.Errorf("OpenProcess(pid=%d): %w", pid, err),
-			resumeACPCommandProcess(pid),
-		)
-	}
-	defer windows.CloseHandle(procHandle)
-	if err := windows.AssignProcessToJobObject(job, procHandle); err != nil {
-		_ = windows.CloseHandle(job)
-		return acpCommandLifecycleHandle{}, errors.Join(
-			fmt.Errorf("AssignProcessToJobObject: %w", err),
-			resumeACPCommandProcess(pid),
-		)
-	}
-	if err := resumeACPCommandProcess(pid); err != nil {
-		_ = windows.CloseHandle(job)
 		return acpCommandLifecycleHandle{}, err
 	}
-	return acpCommandLifecycleHandle{handle: job}, nil
+	return acpCommandLifecycleHandle{job: job}, nil
 }
 
 func releaseACPCommandLifecycle(lifecycle acpCommandLifecycleHandle) {
-	if lifecycle.handle != 0 {
-		_ = windows.CloseHandle(lifecycle.handle)
-	}
-}
-
-func createACPKillOnCloseJob() (windows.Handle, error) {
-	job, err := windows.CreateJobObject(nil, nil)
-	if err != nil {
-		return 0, fmt.Errorf("CreateJobObject: %w", err)
-	}
-	info := windows.JOBOBJECT_EXTENDED_LIMIT_INFORMATION{
-		BasicLimitInformation: windows.JOBOBJECT_BASIC_LIMIT_INFORMATION{
-			LimitFlags: windows.JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE,
-		},
-	}
-	if _, err := windows.SetInformationJobObject(
-		job,
-		windows.JobObjectExtendedLimitInformation,
-		uintptr(unsafe.Pointer(&info)),
-		uint32(unsafe.Sizeof(info)),
-	); err != nil {
-		_ = windows.CloseHandle(job)
-		return 0, fmt.Errorf("SetInformationJobObject: %w", err)
-	}
-	return job, nil
-}
-
-func resumeACPCommandProcess(pid int) error {
-	snapshot, err := windows.CreateToolhelp32Snapshot(windows.TH32CS_SNAPTHREAD, 0)
-	if err != nil {
-		return fmt.Errorf("CreateToolhelp32Snapshot: %w", err)
-	}
-	defer windows.CloseHandle(snapshot)
-
-	entry := windows.ThreadEntry32{Size: uint32(unsafe.Sizeof(windows.ThreadEntry32{}))}
-	if err := windows.Thread32First(snapshot, &entry); err != nil {
-		return fmt.Errorf("Thread32First: %w", err)
-	}
-
-	resumed := 0
-	for {
-		if entry.OwnerProcessID == uint32(pid) {
-			thread, err := windows.OpenThread(windows.THREAD_SUSPEND_RESUME, false, entry.ThreadID)
-			if err != nil {
-				return fmt.Errorf("OpenThread(thread_id=%d): %w", entry.ThreadID, err)
-			}
-			if _, err := windows.ResumeThread(thread); err != nil {
-				_ = windows.CloseHandle(thread)
-				return fmt.Errorf("ResumeThread(thread_id=%d): %w", entry.ThreadID, err)
-			}
-			_ = windows.CloseHandle(thread)
-			resumed++
-		}
-		if err := windows.Thread32Next(snapshot, &entry); err != nil {
-			if errors.Is(err, windows.ERROR_NO_MORE_FILES) {
-				break
-			}
-			return fmt.Errorf("Thread32Next: %w", err)
-		}
-	}
-	if resumed == 0 {
-		return fmt.Errorf("no threads found for pid %d", pid)
-	}
-	return nil
+	_ = lifecycle.job.Close()
 }
 
 func terminateACPProcessGroup(pid int) error {
@@ -136,15 +46,7 @@ func killACPProcessGroup(pid int) error {
 }
 
 func runTaskkill(args ...string) error {
-	output, err := exec.Command("taskkill", args...).CombinedOutput()
-	if err == nil {
-		return nil
-	}
-	msg := strings.TrimSpace(string(output))
-	if msg == "" {
-		return err
-	}
-	return fmt.Errorf("%w: %s", err, msg)
+	return winproc.RunTaskkill(args...)
 }
 
 func acpProcessGroupAlive(pid int) bool {
