+++ apps/backend/internal/agentctl/server/winproc/lifecycle_other.go
@@ -1,0 +1,3 @@
+//go:build !windows
+
+package winproc
